Make configutil tests more standalone, and less repetitive.
